Adding Example 3
accepts any string that starts in the either a 'c' or an 'a' (can start in the middle) if it is a 'c' the read head will move left as long as their is a 'b' to the left and will end on an 'a'. If starts at 'a' then it will move right as long as their is a 'b' and will end on a 'c'
Ex. abbbb>cbbbc  bbbb>abbbbbc